Restucture lib folder into pkg and update cmd to pull from the correct place
