--- lib/logformatter.go
@@ -1,5 +1,0 @@
-package lib
-
-// This file is intentionally left empty.
-// The functionality previously in this file has been merged into logparser.go
-// to avoid duplication and improve code organization.
