+++ pkg/format/container.go
@@ -1,4 +1,4 @@
-package lib
+package format
 
 import (
 	"strings"
