+++ pkg/logging/parser.go
@@ -1,25 +1,23 @@
-package lib
+package logging
 
 import (
-	"bufio"
 	"fmt"
-	"io"
 	"regexp"
 	"strings"
 
 	"github.com/fatih/color"
 )
 
+type LogEntry struct {
+	Level   LogLevel
+	Message string
+}
+
 var logLevelColors = map[LogLevel]*color.Color{
 	DEBUG: color.New(color.FgCyan),
 	INFO:  color.New(color.FgGreen),
 	WARN:  color.New(color.FgYellow),
 	ERROR: color.New(color.FgRed),
-}
-
-func ParseLog(log string) string {
-	entry := ParseLogEntry(log)
-	return logLevelColors[entry.Level].Sprint(log)
 }
 
 func ParseLogEntry(line string) LogEntry {
@@ -53,13 +51,7 @@ func ParseLogLevel(level string) (LogLevel, error) {
 	}
 }
 
-func FilterAndFormatLogs(reader io.Reader, writer io.Writer, filterLevel LogLevel) error {
-	scanner := bufio.NewScanner(reader)
-	for scanner.Scan() {
-		entry := ParseLogEntry(scanner.Text())
-		if entry.Level >= filterLevel {
-			logLevelColors[entry.Level].Fprintf(writer, "%s\n", entry.Message)
-		}
-	}
-	return scanner.Err()
+func ParseLog(log string) string {
+	entry := ParseLogEntry(log)
+	return logLevelColors[entry.Level].Sprint(log)
 }
